pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sundial-airflow-utils"
version = "0.1.0"
description = "Shared Airflow + Astronomer Cosmos utilities for Sundial dbt tenants."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Sundial" }]
license = { text = "Proprietary" }
dependencies = [
    "astronomer-cosmos>=1.12,<2.0",
    "apache-airflow>=2.8",
    "apache-airflow-providers-slack",
    "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/sundialso/sundial-airflow-utils"

[tool.setuptools.packages.find]
include = ["sundial_airflow*"]
